nango dual flow preparatory work (#6346)
* nango dual flow preparatory work

* fix

* fix
